Moving VariableLiveRanges classes outside Compiler class
