+++ src/jit/codegenarmarch.cpp
@@ -2178,7 +2178,7 @@ void CodeGen::genRegCopy(GenTree* treeNode)
 
 #ifdef USING_VARIABLE_LIVE_RANGE
                 // Report the home change for this variable
-                compiler->getVariableLiveKeeper()->siUpdateVariableLiveRange(varDsc, lcl->gtLclNum);
+                varLiveKeeper->siUpdateVariableLiveRange(varDsc, lcl->gtLclNum);
 #endif // USING_VARIABLE_LIVE_RANGE
 
                 // The new location is going live
