+++ src/jit/treelifeupdater.cpp
@@ -248,7 +248,7 @@ void TreeLifeUpdater<ForCodeGen>::UpdateLifeVar(GenTree* tree)
 
 #ifdef USING_VARIABLE_LIVE_RANGE
             // For each of the LclVarDsc that are reporting change, variable or fields
-            compiler->getVariableLiveKeeper()->siStartOrCloseVariableLiveRanges(varDeltaSet, isBorn, isDying);
+            compiler->codeGen->getVariableLiveKeeper()->siStartOrCloseVariableLiveRanges(varDeltaSet, isBorn, isDying);
 #endif // USING_VARIABLE_LIVE_RANGE
 
 #ifdef USING_SCOPE_INFO
